CR fixes #52
