+++ dontforget/app.py
@@ -18,7 +18,6 @@ def create_app(config_object=ProdConfig):
     register_extensions(app)
     register_blueprints(app)
     register_errorhandlers(app)
-
     return app
