+++ dontforget/public/forms.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
 """Public forms."""
-from flask_wtf import Form
+from flask_wtf import FlaskForm
 from wtforms import PasswordField, StringField
 from wtforms.validators import DataRequired
 
 from dontforget.user.models import User
 
 
-class LoginForm(Form):
+class LoginForm(FlaskForm):
     """Login form."""
 
     username = StringField('Username', validators=[DataRequired()])
