--- helpers/__init__.py
@@ -1,1 +1,0 @@
-"""Helper methods and extensions to some modules."""
